/pkcs7_x509.c functions implemented, others not
